fix(calendar): setting timezone to 'Europe/Brussels'
This changes the default timezone from 'locale' to 'Europe/Brussels'. We could let the user choose but in my opinion this will add verbosity to the web interface and ADE also uses the same timezone, regardless where you are on the planet.

This fixes the problem when you are in a different timezone and classes are outside exposed hours (before 8 am or after 10 pm).