fix(bug): enable lazy course update
This enables to update courses only if the request succeeds. If the request fails and the course was already present, then we use previous version.

This should help lowering the number of error messages we have. With this, I propose to reduce the interval time to 30 minutes, since any error will add another 30 minutes.

This feature can also be activated for other requests, but will required quite some work and I am not sure it's that important for the moment :'-)