feat: add some randomness to compute
Adding randomness to compute allows for events to appear (hopefully) equally often. Currently, this was not possible because events at the end of the week were less likely to be taken.

Using a seed is important so that all weeks are shuffled the same way.

Using `sort=False` reduces computational time, since we do not need sorted entries.